+++ browser/js/twitch/background.js
@@ -19,7 +19,7 @@ function getVideoMaxResolution(resolutions){
     }
 }
 
-async function getTargetInfo(url, need_stream = true){
+async function getTargetInfo(url, need_playable = true){
     let url_info = parseUrl(url);
     let target_info = {};
     if(url_info.paths[0] == 'videos'){
@@ -34,14 +34,41 @@ async function getTargetInfo(url, need_stream = true){
             let response_json = await response.json();
             if(response_json){
                 target_info.channel_name = response_json.channel.name;
-                target_info.channel_id = response_json.channel._id;
-                target_info.video.id = response_json._id;
+                target_info.channel_id = parseInt(response_json.channel._id);
+                target_info.video.id = parseInt(response_json._id);
                 target_info.video.resolution = getVideoMaxResolution(Object.values(response_json.resolutions));
+            }
+        }
+    }else if(url_info.paths[1] == 'clip' || url_info.domains[2] == 'clips'){
+        let slug;
+        if(url_info.paths[1] == 'clip'){
+            slug = url_info.paths[2];
+        }else if(url_info.domains[2] == 'clips'){
+            if(url_info.paths[0] == 'embed'){
+                slug = url_info.searchParams.get('clip');
+            }else{
+                slug = url_info.paths[0];
+            }
+        }
+        if(slug){
+            let response = await fetch(TWITCH_KRAKEN + '/clips/' + slug, TWITCH_FETCH_INIT);
+            if(response && response.status == 200){
+                let response_json = await response.json();
+                if(response_json){
+                    target_info.clip = {};
+                    target_info.clip.slug = response_json.slug;
+                    target_info.channel_name = response_json.broadcaster.name;
+                    target_info.channel_id = parseInt(response_json.broadcaster.id);
+                    if(response_json.vod && need_playable){
+                        let vod_target_info = await getTargetInfo(response_json.vod.url);
+                        target_info.video = vod_target_info.video;
+                    }
+                }
             }
         }
     }else{
         target_info.channel_name = url_info.paths[0];
-        if(need_stream){
+        if(need_playable){
             let response = await fetch(TWITCH_KRAKEN + '/users' + '?login=' + target_info.channel_name, TWITCH_FETCH_INIT);
             if(response && response.status == 200){
                 let response_json = await response.json();
@@ -58,7 +85,7 @@ async function getTargetInfo(url, need_stream = true){
                             target_info.stream.resolution = {};
                             target_info.stream.resolution.height = response_json.stream.video_height;
                             target_info.stream.resolution.width = response_json.stream.video_height * (16/9);
-                            let stream_id = response_json.stream._id;
+                            let stream_id = parseInt(response_json.stream._id);
                             let response = await fetch(TWITCH_KRAKEN + '/channels' + '/' + target_info.channel_id + '/videos' + '?limit=4&broadcast_type=archive', TWITCH_FETCH_INIT);
                             if(response && response.status == 200){
                                 let response_json = await response.json();
@@ -108,7 +135,14 @@ async function openPlayer(target_info){
     let url = new URL('https://player.twitch.tv/');
     url.searchParams.append('volume', '1');
     let resolution;
-    if(target_info.video){
+    if(target_info.clip){
+        url.hostname = 'clips.twitch.tv';
+        url.pathname = 'embed';
+        url.searchParams.append('clip', target_info.clip.slug);
+        if(target_info.video && target_info.video.resolution){
+            resolution = target_info.video.resolution;
+        }
+    }else if(target_info.video){
         url.searchParams.append('video', target_info.video.id);
         if(target_info.video.time){
             url.searchParams.append('time', target_info.video.time);