+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Added
+ - Clips popout support
+
 ## 1.0.1 - 2019-09-29
 ### Changed
  - Chat header link adder updated to new Twitch style